refactor optimize executor
