+++ optimizer/src/main/java/com/netease/arctic/optimizer/operator/executor/IcebergMajorExecutor.java
@@ -20,47 +20,27 @@ package com.netease.arctic.optimizer.operator.executor;
 
 import com.netease.arctic.optimizer.OptimizerConfig;
 import com.netease.arctic.table.ArcticTable;
-import org.apache.iceberg.ContentFile;
-import org.apache.iceberg.relocated.com.google.common.collect.Lists;
+import org.apache.iceberg.DataFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class IcebergMinorExecutor<F extends ContentFile<F>> extends BaseExecutor<F> {
-
+public class IcebergMajorExecutor extends BaseExecutor<DataFile> {
   private static final Logger LOG = LoggerFactory.getLogger(IcebergMajorExecutor.class);
 
-  public IcebergMinorExecutor(NodeTask nodeTask, ArcticTable table, long startTime, OptimizerConfig config) {
+  public IcebergMajorExecutor(NodeTask nodeTask, ArcticTable table, long startTime, OptimizerConfig config) {
     super(nodeTask, table, startTime, config);
   }
 
   @Override
-  public OptimizeTaskResult<F> execute() throws Exception {
-    LOG.info("Start processing iceberg table minor optimize task: {}", task);
+  public OptimizeTaskResult<DataFile> execute() throws Exception {
+    LOG.info("Start processing iceberg major optimize task: {}", task);
 
-    Iterable<F> targetFiles = table.io().doAs(() -> {
-      if (task.icebergDataFiles().size() > 0) {
-        // optimize iceberg delete files.
-        return optimizeDeleteFiles();
-
-      } else if (task.icebergSmallDataFiles().size() > 0) {
-        // optimize iceberg small data files.
-        return optimizeSmallDataFiles();
-      }
-      return Lists.newArrayList();
+    Iterable<DataFile> targetFiles = table.io().doAs(() -> {
+      //TODO
+      return null;
     });
 
-
     return buildOptimizeResult(targetFiles);
-  }
-
-  private Iterable<F> optimizeDeleteFiles() throws Exception {
-    // TODO
-    return null;
-  }
-
-  private Iterable<F> optimizeSmallDataFiles() throws Exception {
-    // TODO
-    return null;
   }
 
   @Override
